moved previous sox compatibility into functional
