--- docs/source/functional_sox_compatibility.rst
@@ -1,24 +1,0 @@
-.. role:: hidden
-    :class: hidden-section
-
-torchaudio.functional_sox_compatibility
-=======================================
-
-.. currentmodule:: torchaudio.functional_sox_compatibility
-
-SoX compatibility functions
-
-:hidden:`biquad`
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: biquad
-
-:hidden:`highpass_biquad`
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: highpass_biquad
-
-:hidden:`lowpass_biquad`
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-
-.. autofunction:: lowpass_biquad
