+++ docs/source/index.rst
@@ -13,7 +13,6 @@ The :mod:`torchaudio` package consists of I/O, popular datasets and common audio
    kaldi_io
    transforms
    functional
-   functional_sox_compatibility
 
 .. automodule:: torchaudio
    :members:
